test(cli): move Ollama-dependent memory tests behind the integration tag
TestRunMemoryMine_ProjectFiles, TestRunMemoryMine_Conversations and
TestRunMemoryStatus_WithDB need a live Ollama daemon and an embedding
model, so they fail on any host without one. Move them out of
cli_coverage_test.go into memory_mine_integration_test.go behind
//go:build integration; the default `go test` build no longer compiles
them. Run with `go test -tags integration ./internal/cli/...`.

Also silence SA1019 in the MCP everything server with a file-scoped
lint directive. The server deliberately exercises the deprecated
SEP-2577 roots/sampling/logging APIs, and the .golangci.yml `^hack/`
exclusion misses them whenever golangci-lint serves a cached result
recorded under a different worktree's path — the pattern is anchored,
the printed path is not. A file-scoped directive is path-independent.
