+++ internal/cli/cli_coverage_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/dimetron/pi-go/internal/logger"
 	"github.com/dimetron/pi-go/internal/lsp"
 	"github.com/dimetron/pi-go/internal/memory"
-	"github.com/dimetron/pi-go/internal/palace"
 	"github.com/dimetron/pi-go/internal/subagent"
 	"github.com/dimetron/pi-go/internal/tools"
 	"github.com/dimetron/pi-go/internal/webserver"
@@ -665,78 +664,6 @@ func TestFindMemoryDB_EmptyProject(t *testing.T) {
 }
 
 // -----------------------------------------------------------------------
-// runMemoryMine — project files
-// -----------------------------------------------------------------------
-
-// skipWithoutEmbedder skips tests that need a real embedding backend.
-//
-// These are integration tests wearing unit-test clothing: they call the same
-// code paths `pi memory mine` does, which embed for real. With the default
-// config that means a local Ollama daemon, and without one they either fail
-// outright ("cannot reach daemon at http://localhost:11434") or — worse — fall
-// through to downloading a ~100 MB model, which blew past the 10-minute test
-// timeout and took the whole package down with a panic rather than a failure.
-//
-// Probing is cheap: EmbedderAvailability is a loopback request with a 3-second
-// deadline. So this skips in CI, where no daemon runs, and still exercises the
-// code locally where one does — which is better than a build tag that would
-// make these dead weight everywhere.
-func skipWithoutEmbedder(t *testing.T) {
-	t.Helper()
-	cfg := palace.DefaultConfig()
-	cfg.ModelPath = defaultPalaceModelPath()
-	if err := palace.EmbedderAvailability(cfg); err != nil {
-		t.Skipf("no embedding backend available: %v", err)
-	}
-}
-
-func TestRunMemoryMine_ProjectFiles(t *testing.T) {
-	// Skip race detection due to race in third-party go-huggingface library
-	// (hugot.DownloadModel → gomlx/go-huggingface hub.(*Repo).DownloadFiles).
-	if raceEnabled {
-		t.Skip("skipping: go-huggingface has race condition")
-	}
-	skipWithoutEmbedder(t)
-	dir := t.TempDir()
-	os.WriteFile(filepath.Join(dir, "main.go"),
-		[]byte("package main\n\nfunc main() { println(\"hello world test content for chunk threshold minimum\") }"), 0o644)
-
-	output := captureStdout(t, func() {
-		err := runMemoryMine(dir, "testproject", false)
-		if err != nil {
-			t.Fatalf("runMemoryMine: %v", err)
-		}
-	})
-	if output == "" {
-		t.Error("expected mining output")
-	}
-}
-
-func TestRunMemoryMine_Conversations(t *testing.T) {
-	// Skip race detection due to race in third-party go-huggingface library
-	// (hugot.DownloadModel → gomlx/go-huggingface hub.(*Repo).DownloadFiles).
-	if raceEnabled {
-		t.Skip("skipping: go-huggingface has race condition")
-	}
-	skipWithoutEmbedder(t)
-	dir := t.TempDir()
-	os.WriteFile(filepath.Join(dir, "chat.jsonl"),
-		[]byte(`{"role":"user","content":"question"}
-{"role":"assistant","content":"answer"}
-`), 0o644)
-
-	output := captureStdout(t, func() {
-		err := runMemoryMine(dir, "testconv", true)
-		if err != nil {
-			t.Fatalf("runMemoryMine convos: %v", err)
-		}
-	})
-	if output == "" {
-		t.Error("expected mining output")
-	}
-}
-
-// -----------------------------------------------------------------------
 // runMemoryModelStatus
 // -----------------------------------------------------------------------
 
@@ -794,38 +721,6 @@ func TestRunMemoryStatus_NoDB(t *testing.T) {
 	})
 	if output == "" {
 		t.Error("expected 'no palace' message")
-	}
-}
-
-func TestRunMemoryStatus_WithDB(t *testing.T) {
-	// AddDrawer embeds its content, so this needs a backend like the mine tests
-	// do. This is the test that hung for 10 minutes and panicked the package.
-	skipWithoutEmbedder(t)
-
-	dir := t.TempDir()
-	dbPath := filepath.Join(dir, "palace.db")
-
-	// Create a real palace with some data
-	p, err := palace.New(palace.WithDBPath(dbPath))
-	if err != nil {
-		t.Fatalf("palace.New: %v", err)
-	}
-	p.AddDrawer(context.Background(), palace.DrawerInput{
-		Wing: "test", Room: "api", Content: "test content",
-	})
-	p.KGAdd(context.Background(), palace.TripleInput{
-		Subject: "Alice", Predicate: "works_on", Object: "api",
-	})
-	p.Close()
-
-	output := captureStdout(t, func() {
-		err := runMemoryStatus(dbPath)
-		if err != nil {
-			t.Fatalf("runMemoryStatus: %v", err)
-		}
-	})
-	if output == "" {
-		t.Error("expected status output")
 	}
 }
